# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /academicDegrees:
    get:
      tags:
        - academic-degrees-controller
      summary: Отобразить все учёные степени
      operationId: findAllAcademicDegrees
      responses:
        200:
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AcademicDegreesResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'

    post:
      tags:
        - academic-degrees-controller
      summary: Создать новую учёную степень
      operationId: createAcademicDegree
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcademicDegreesCreateRequest'
      responses:
        201:
          description: Ученая степень успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicDegreesResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'
  /academicDegrees/{id}:
    put:
      tags:
        - academic-degrees-controller
      summary: Изменить информацию об ученой степени  по ID
      operationId: updateAcademicDegree
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcademicDegreesUpdateRequest'
      responses:
        200:
          description: Ученая степень успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicDegreesResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
  /academicTitles:
    get:
      tags:
        - academic-titles-controller
      summary: Отобразить все учёные звания
      operationId: findAllAcademicTitles
      responses:
        200:
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AcademicTitlesResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'

    post:
      tags:
        - academic-titles-controller
      summary: Создать новое учёное звание
      operationId: createAcademicTitle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcademicTitlesCreateRequest'
      responses:
        201:
          description: Ученое звание успешно создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicTitlesResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'
  /academicTitles/{id}:
    put:
      tags:
        - academic-titles-controller
      summary: Изменить информацию об ученом звании по ID
      operationId: updateAcademicTitle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcademicTitlesUpdateRequest'
      responses:
        200:
          description: Ученое звание успешно обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicTitlesResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
  /educationSubject:
    get:
      tags:
        - education-subject-controller
      summary: Отобразить все возможные научные направления
      operationId: findAllEducationSubjects
      responses:
        200:
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EducationSubjectResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'

    post:
      tags:
        - education-subject-controller
      summary: Создать новое научное направление
      operationId: createEducationSubject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EducationSubjectCreateRequest'
      responses:
        201:
          description: Научное направление успешно создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EducationSubjectResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'
  /educationSubject/{id}:
    put:
      tags:
        - education-subject-controller
      summary: Изменить информацию о научном направлении по ID
      operationId: updateEducationSubject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EducationSubjectUpdateRequest'
      responses:
        200:
          description: Научное направление успешно обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EducationSubjectResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
  /achievements:
    get:
      tags:
        - achievements-controller
      summary: Отобразить все достижения
      operationId: findAllAchievements
      responses:
        200:
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AchievementResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'

    post:
      tags:
        - achievements-controller
      summary: Создать новую запись про достижение
      operationId: createAchievement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AchievementCreateRequest'
      responses:
        201:
          description: Достижение успешно создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AchievementResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'
  /achievements/{personId}:
    get:
      tags:
        - achievements-controller
      summary: Найти достижение по ID человека
      operationId: findByPersonId
      parameters:
        - name: personId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        200:
          description: Список достижений человека
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AchievementResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
  /achievements/{id}:
    put:
      tags:
        - achievements-controller
      summary: Изменить запись заслуги по ID
      operationId: updateAchievement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AchievementUpdateRequest'
      responses:
        200:
          description: Достижение успешно обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AchievementResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
...
  /jobTitle:
    get:
      tags:
        - job-title-controller
      summary: Отобразить все возможные должности
      operationId: findAllJobTitles
      responses:
        200:
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobTitleResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'

    post:
      tags:
        - job-title-controller
      summary: Создать новую должность
      operationId: createJobTitle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobTitleCreateRequest'
      responses:
        201:
          description: Должность успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobTitleResponse'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'
  /jobTitle/{id}:
    put:
      tags:
        - job-title-controller
      summary: Изменить информацию о должности
      operationId: updateJobTitle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobTitleUpdateRequest'
      responses:
        200:
          description: Должность успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobTitleResponse'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
...
components:
  schemas:
    AcademicDegreesCreateRequest:
      type: object
      properties:
        name:
          type: string
          description: Название ученой степени
          example: "доктор наук"
        shortName:
          type: string
          description: Сокращенное название
          example: "д.н."
      required:
        - name
    AcademicDegreesUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/AcademicDegreesCreateRequest'
        - type: object
          properties:
            display:
              type: boolean
              default: true
              nullable: true
    AcademicDegreesResponse:
      allOf:
        - $ref: '#/components/schemas/AcademicDegreesCreateRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              description: ID учёной степени
          required:
            - id
    AcademicTitlesCreateRequest:
      type: object
      properties:
        name:
          type: string
          description: Название ученого звания
          example: "профессор"
        shortName:
          type: string
          description: Сокращенное название
          example: "проф."
      required:
        - name
    AcademicTitlesUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/AcademicTitlesCreateRequest'
        - type: object
          properties:
            display:
              type: boolean
              default: true
              nullable: true
    AcademicTitlesResponse:
      allOf:
        - $ref: '#/components/schemas/AcademicTitlesCreateRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              description: ID учёного звания
          required:
            - id
    Achievement:
      type: object
      properties:
        title:
          type: string
          description: Название достижения
          example: "Нобелевская премия"
        year:
          type: integer
          format: int32
          description: Год получения достижения
          example: 2020
        description:
          type: string
          description: Описание достижения
          example: "За открытие в области физики"
      required:
        - title
    AchievementResponse:
      allOf:
        - $ref: '#/components/schemas/Achievement'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              description: ID достижения
          required:
            - id
    AchievementCreateRequest:
      allOf:
        - $ref: '#/components/schemas/Achievement'
        - type: object
          properties:
            personId:
              type: integer
              format: int64
              description: Идентификатор человека, связанного с достижением
              example: 5
          required:
            - personId
    AchievementUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/AchievementCreateRequest'
        - type: object
          properties:
            display:
              type: boolean
              default: true
              nullable: true
    EducationSubjectCreateRequest:
      type: object
      properties:
        name:
          type: string
          description: Название учебного предмета
          example: "экономика"
        shortName:
          type: string
          description: Сокращенное название предмета
          example: "эконом."
      required:
        - name
    EducationSubjectUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/EducationSubjectCreateRequest'
        - type: object
          properties:
            display:
              type: boolean
              default: true
              nullable: true
    EducationSubjectResponse:
      allOf:
        - $ref: '#/components/schemas/EducationSubjectCreateRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              description: ID научного направления
          required:
            - id
    JobTitleCreateRequest:
      type: object
      properties:
        name:
          type: string
          description: Название должности
          example: "директор"
      required:
        - name
    JobTitleUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/JobTitleCreateRequest'
        - type: object
          properties:
            display:
              type: boolean
              default: true
              nullable: true
    JobTitleResponse:
      allOf:
        - $ref: '#/components/schemas/JobTitleCreateRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              description: ID должности
          required:
            - id
...
  responses:
    400:
      description: Некорректный запрос
    404:
      description: Ресурс не найден
    500:
      description: Внутренняя ошибка сервера